# Change CPU geometric_kernel from double to double/float