I just don't understand why on CPU we use `double` but on CUDA we use `scalar_t`/`accscalar_t`.

I didn't see any performance improvements(or degradations) but I just want CPU kernels to be consistent with CUDA code and use `scalar_t`/`accscalar_t`.

```
import torch

t = torch.empty(1000, dtype=torch.double, device='cpu')
timeit t.geometric_(0.5)
BEFORE: 59.4 µs ± 288 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
AFTER : 59.3 µs ± 265 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)


t = torch.empty(1000000, dtype=torch.double, device='cpu')
timeit t.geometric_(0.5)
BEFORE: 57.8 ms ± 317 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
AFTER : 57.8 ms ± 154 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)


t = torch.empty(1000, dtype=torch.float, device='cpu')
timeit t.geometric_(0.5)
BEFORE: 60 µs ± 167 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
AFTER : 58.9 µs ± 137 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)


t = torch.empty(1000000, dtype=torch.float, device='cpu')
timeit t.geometric_(0.5)
BEFORE: 58.7 ms ± 238 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
AFTER : 57.5 ms ± 182 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)


t = torch.empty(1000, dtype=torch.half, device='cpu')
timeit t.geometric_(0.5)
BEFORE: 67.2 µs ± 750 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
AFTER : 66.1 µs ± 513 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)


t = torch.empty(1000000, dtype=torch.half, device='cpu')
timeit t.geometric_(0.5)
BEFORE: 65.6 ms ± 187 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
AFTER : 64.6 ms ± 464 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Stack from [ghstack](https://github.com/ezyang/ghstack):
* #36633 CSPRNG PoC
* #38406 Change CPU log_normal_kernel from double to double/float
* #38405 Change CPU cauchy_kernel from double to double/float
* #38404 Change CPU exponential_kernel from double to double/float
* **#38371 Change CPU geometric_kernel from double to double/float**
* #38427 Add bfloat16 to CPU cauchy_kernel, log_normal_kernel, exponential_kernel
* #37984 RNG infrastructure improvements